Refactor imports in manager.ts and manager-helpers.tsx for consistency
